testing

+++ ScaFFold/worker.py
@@ -247,7 +247,11 @@ def main(kwargs_dict: dict = {}):
     )
 
     # solve hang?
-    dist.barrier()
+    if os.getenv("SKIP_DIST_BARRIERS") != "1":
+        torch.cuda.synchronize()
+        print(f"Done cuda sync rank {rank}")
+        dist.barrier()
+        print(f"Done barrier rank {rank}")
 
     #
     # Generate plots
@@ -258,7 +262,9 @@ def main(kwargs_dict: dict = {}):
         standard_viz.main(config)
         end_code_region("generate_figures")
 
-    dist.barrier()
-    dist.destroy_process_group()
+    if os.getenv("SKIP_DIST_BARRIERS") != "1":
+        dist.barrier()
+        print(f"Done barrier rank {rank}")
+        dist.destroy_process_group()
 
     return 0
